Resolve import cycle
