+++ backend/sendmessage/src/reset-votes.ts
@@ -2,7 +2,7 @@ import { getConnection } from '../../shared/database/getConnection';
 import { getGroup } from '../../shared/database/getGroup';
 import { resetGroupVotes } from '../../shared/database/resetGroupVotes';
 import { ConfigWithHandler } from '../../shared/types';
-import { broadcastState } from '../../shared/broadcast-state';
+import { broadcastState } from '../../shared/actions';
 
 export const resetVotes = async (config: ConfigWithHandler) => {
   const connectionItem = await getConnection(config);
