+++ backend/sendmessage/src/set-scale.ts
@@ -3,7 +3,7 @@ import { getConnection } from '../../shared/database/getConnection';
 import { getGroup } from '../../shared/database/getGroup';
 import { resetGroupVotes } from '../../shared/database/resetGroupVotes';
 import { ConfigWithHandler } from '../../shared/types';
-import { broadcastState } from '../../shared/broadcast-state';
+import { broadcastState } from '../../shared/actions';
 
 export const setScale = async (scale: CardValue[], config: ConfigWithHandler): Promise<void> => {
   const connectionItem = await getConnection(config);
